Rename Rock_vs_Mine.ipynb to Rock_vs_Mine_Prediction.ipynb
In modern warfare and security, Submarines play a crucial role and so they face severe threats from mines set in deep waters. To avoid casualties and ensure safe voyage, Submarines use the SONAR (Sound Navigation and Ranging) that is helpful for exploring and mapping the ocean by using sound waves.

From the data source, this project aims to predict whether the underwater obstacle is a Mine or just a rock.
The Prediction is done by Logistic Regression model.